+++ Documentation/documentation.sh
@@ -7,15 +7,9 @@ PROJECT_PATH=com.chartboost.core.canary     # The unity project where this packa
 # Generate .csproj files
 "$UNITY_EXE_PATH" -quit -batchmode -nographics -projectPath $PROJECT_PATH -executeMethod Packages.Rider.Editor.RiderScriptEditor.SyncSolution
 
-# Copy files
-echo "Copying files..."
-cp com.chartboost.core/README.md com.chartboost.core/Documentation/index.md
-cp com.chartboost.core/CONTRIBUTING.md com.chartboost.core/Documentation/CONTRIBUTING.md
-cp com.chartboost.core/LICENSE.md com.chartboost.core/Documentation/LICENSE.md
-
-# Update links in index.md
-echo "Updating links..."
-sed -i '' 's|\(Documentation/\)\([^)]*\)|\2|g' com.chartboost.core/Documentation/index.md
+# Generate index.md
+chmod +x com.chartboost.unity.ci/scripts~/generate-docfx-index.sh
+com.chartboost.unity.ci/scripts~/generate-docfx-index.sh "com.chartboost.core/Documentation" "Chartboost.Core.html"
 
 # Generate Documentation
 echo "Generating documentation.."