+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file using the standards as defined at [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
+### Version 1.0.2 *(2024-09-19)*
+Improvements:
+- Modified `NativeModule` internal class to expand API flexibility of wrapped native modules.
+
+Bug Fixes:
+- `ConsentKey` predefined values are in sync with native values.
+
 ### Version 1.0.1 *(2024-08-29)*
 Fixed:
 - Skipped module identifiers not forwarded to iOS Core SDK.